fix: sales channel admin component inheritance
