--- src/Resources/app/administration/src/module/swag-paypal-agent-commerce/extension/sw-sales-channel-modal/index.ts
@@ -1,28 +1,0 @@
-import { SwagPayPalDefaults } from "SwagPayPal/defaults";
-
-export default Shopware.Component.wrapComponentConfig<{ onCloseModal: () => void }>({
-    methods: {
-        onAddChannel(id: string | null) {
-            if (id === SwagPayPalDefaults.agentCommerceTypeId) {
-                this.onCloseModal();
-
-                if (id) {
-                    this.$router.push({
-                        name: 'swag.paypal.agent.commerce.create',
-                        params: { typeId: id },
-                    });
-                }
-            }
-
-            this.$super('onAddChannel', id);
-        },
-
-        isProductComparisonSalesChannelType(salesChannelTypeId: string): boolean {
-            if (salesChannelTypeId === SwagPayPalDefaults.agentCommerceTypeId) {
-                return true;
-            }
-
-            return this.$super('isProductComparisonSalesChannelType', salesChannelTypeId) as boolean;
-        },
-    },
-});
