+++ src/Resources/app/administration/src/module/swag-paypal-agent-commerce/static/product-export/header.txt
@@ -1,27 +1,16 @@
-id,
-title,
-description,
-link,
-image_link,
-additional_image_link,
-price,
-sale_price,
-availability,
-condition,
-brand,
-gtin,
-mpn,
-google_product_category,
-product_type,
-shipping,
-shipping_weight,
-shipping_label,
-item_group_id,
-color,
-size,
-material,
-pattern,
-gender,
-age_group,
-size_type,
-size_system
+"id",
+"item_group_id",
+"title",
+"link",
+"mpn",
+"brand",
+"gtin",
+"sale_price",
+"price",
+"availability",
+"availability_date",
+"google_product_category",
+"product_type",
+"image_link",
+"additional_image_link",
+"description"
